+++ src/general.h
@@ -2,9 +2,9 @@
 #define GENERAL_H
 #include "math.h"
 
-float random();
-float gaussiana(float,float);
-float norma2(float*, float*, int);
-float delta_x(float, float, float);
+double random();
+double gaussiana(double,double);
+double norma2(double*, double*, int);
+double delta_x(double, double, double);
 
 #endif
